7685 Dismiss connection modal when cancel complete
Previously, if you cancelled a card reader search and then started a new one too quickly, the SDK would not be ready to perform further actions and would show an error.

Cancellation of SDK actions is asynchronous in its own right. We had multiple calls to dismiss fired by different parts of the modal, so in some cases we risked dismissing extra alerts if they were presented.

This removes and moves dismiss calls, to handle them after the cancellation and any completion handlers have run.