Hide `Mark Order Complete` on unpaid CoD orders
The main CTA for a cash on delivery button should be `Collect Payment`, which showed before but was below the fold, with `Mark Order Complete` visible at the top. If tapped, this would have marked the order as paid even if the merchant had not recieved payment, and prevented them taking a card payment for the order.

This change prevents the `markCompleteButton` from being added on orders which are still `isEligibleForPayment`, the criterion used for showing the `Collect Payment` button.