Build Experiments framework using app extension safe APIs
The build process has been warning us about Experiments, Networking, and
WooFoundation not being "safe" to link in App Extension targets.

Setting this build setting to true seems to address the issue.

The warning was like:

```
warning build:
Linking against a dylib which is not safe for use in application extensions:
/path/to/DerivedData/WooCommerce/Build/Products/Debug-iphonesimulator/Networking.framework/Networking
```