Show plugin updates after webview dismissed
We had to remove the protocol for the viewmodel, or use type erasure. It was simpler to use a non-specific viewmodel which could be configured with the plugin name, and just set the properties in the Previews.

This is a bit of a shame because the protocol approach was nicer, but Swift didn’t let us use it with ObservableObject, and type erasure has quite a lot of overhead.