Remove mapUniqueCustomer in favor of its Remote
When Unit Testing the previous retrieveCustomerByName(), we uncovered an error where the resultant array of entities was not only the used search term, but all entities from the endpoint. This is because the Mapper calls only the WCAnalyticsCustomerMapper.map() function, and not other methods. In this case, we remove the mapUniqueCustomer() from the Mapper, and add the logic to WCAnalyticsCustomerRemote instead.