Sync plugins when settings is opened
This ensures that we have the latest plugin data to show whether the WooCommerce plugin has available updates. If we don’t sync it when settings is opened, we may be relying on very old stored data, depending on the user’s usage pattern for the app.

Previously, the plugins were only synced on login/fresh launch, or for a few IPP-related actions.