Make keyword non-optional. Simplify test helper
Since we’re making the “keyword” attribute non-optional, we’re also providing a default value to the CustomerSearchResult entity so we can migrate between data models without issues, as any existing CustomerSearchResult would need to have a “keyword” set to it when we’re migrating data sets. With this we also altered the 74_to_75 tests to check migrated entities between data models, rather than new entities in the new data model.