7908 Add Swift Concurrency `enqueue` to Remote
N.B. This does the same as the existing `enqueue(request: mapper:) -> Result` method, so it could be a simple wrapper. I opted against that in case we decide to convert everything to use async/await… but perhaps that’s a bit optimistic, so it would also make sense to turn it into a simple wrapper.