Remove local StoreKit configuration
Now that we have the products set up in App Store connect and are ready for backend testing, we don't need the local configuration.
Removing it is best, as once you use it, the app receipt will always be on the "xcode" environment until you reinstall the app, and receipt validation won't work.