7916 Use JITM view model with existing UI
`AnnouncementCardViewModelProtocol` leaves it open for the other Product onboarding work to pass its own VM, or a `FeatureAnnouncementCardViewModel`, whichever makes most sense.

There can be further refactoring of this, and it would probably be useful, but I’ll do some in a future PR. In particular, the dismiss handling is a bit specific, and might benefit from being extracted.