8222 Make Preparing Reader background one color
On the Preparing Reader modal, the borders of the hosting controller view can be bigger than the spinner view itself – this is the case on iPad.

It was observed that in dark mode, this shows with two black bars, above and below the spinner.

This changes the background specifier to be on the hosting controller’s view, not the spinner contained within it, which corrects the background color issue.