Continue to use bluetoothScan from manage reader
This corrects an issue where if the feature flag was on, we would search for localMobile readers in the Manage card reader screen. That was not intended, as there’s nothing to manage about localMobile readers.

In future, the Manage card readers screen will start to use the non-legacy connection controller; by doing so, it’ll always get `bluetoothScan`, same as in this commit.