8082 Show specific configuration progress alerts
The iPhone’s built in card reader may need to be configured before it can be used. In this case, Stripe will call our LocalMobileConnectionDelegate with progress of the configuration, in the same way as it does for bluetooth reader software updates.

This commit duplicates and updates the bluetooth reader software update alert for use with configuration updates.