8289 Show Processing after Collect completes in new flow
Previously, we only showed `Processing payment` once we had an authorized (but not captured) payment intent back from Stripe Terminal, and we had sent that to WCPay/Stripe Gateway for capture.

This commit updates the new Payment flow to show the `Processing` message as soon as we have collected the card details from the reader, i.e. when we are about to start the on-reader processing step.

This avoids a gap where the phone displayed `Tap card to pay` after the payment card had actually been read.

This commit only updates the new payment flow