8085 Handle cancelation from a card reader
Previously (prior to  5233d70b ) we didn’t dismiss any alerts when cancel was tapped on the reader (or built in reader modal) because we ignored the error.

This corrects the comments, warnings, and errors produced (in the StripeCardReaderService) now that we don’t ignore the error any more.

It also stops showing an alert (with a try again button) when the error is recieved in the CollectOrderPaymentUseCase, and instead dismisses the payment alerts.

Now, in the new flow, we handle cancelations from card readers and from buttons in the app equally: we dismiss the alerts and track the cancelation events.