8085 Inform the merchant about on-reader cancels
If the user cancels the payment on the reader, it can help the merchant to specifically notify them that the payment has not been taken, and they still need to capture payment.

We do not do this with the built-in reader because it’s clear when everything happens on the same device.