Update a `var` from `public` to `internal` in `MockStorageManager`
Using the strictest access level possible ensures we as little
information as possible, which improves encapsulation and information
hiding.

I noticed this code because of the `TODO` flagging it and I was looking
through the `TODO`s as part of day 2 task of the Code Quality Challenge
"Nuke TODO comments".
See https://tuple.app/code-quality-challenge