8355 Hide IPP image and progress views in landscape
In the past, we hid the image when the IPP payment flow was used in landscape.

With the addition of the progress view in #8079 c: 83f1d4cf0b1134deb90a049ea5dc2dde59a51f96  we added another place where the image view could be shown or hidden (as the progress view replaces the image.)

That introduced a bug where images would be shown in landscape, breaking the rest of the IPP layout.

This commit moves all the `isHidden` changes to the `resetHeightAndWidth`, which is responsible for whether and which view should be shown.