8355 Consistent button size on orientation change
Previously, if you had the phone in landscape when the “Payment Successful” modal was shown, the two primary buttons would be displayed side by side at different sizes. Changing to portrait and back to landscape resulted in two different sizes showing, which felt strange.

Using the `fillProportionally` option resolves this and leads to a more consistent view in this edge case.