8314 Use checkDeviceSupport to show reader choice
Previously the logic for when we show the `Tap to Pay on iPhone` or `Bluetooth reader` screen was in the preflight controller, using Apple’s API’s. Stripe’s support is more restricted than Apple’s, and has been exposed via the CardPresentPaymentsStore. This commit updates the PreflightController to use the exposed Stripe check.