8664 Disconnect built-in reader when Manage tapped
There’s nothing to manage or view about the built in reader, and from a user’s perspective there isn’t a reader at all, it’s just part of their phone.

Since we can’t connect to the Bluetooth reader until the built in reader is disconnected, we just do that automatically for the user here.

This saves us updating the display of the connected reader to be suitable for the built in reader too.

We can revisit this in future, but for M1 it makes sense.