Update Processing title to show payment was started
The title for the Processing Payment modal view was still `Collect payment from {Name}`. This could be percieved as us telling the merchant that [they need to]… collect payment from {name}, which isn’t the case.

This updates the alert to say `Processing payment from {name}` once the payment method details were collected and processing is ongoing.