Assign the survey inside a completion handler
This fixes a bug where we assign the survey before the view model has had time to calculate the correct survey, hence always returning the case for “No IPP transactions”, as the results controller has zero fetchecObjects at the time of assigning the variable. By assigning the survey inside a completion handler, we’re sure that is only assigned when the fetch has completed.