Use PaymentGateway from onboarding during payment
Previously we fetched and passed around the Payment Gateway independently of running onboarding. This meant that when we moved onboarding to be handled in Preflight, the PaymentGatewayAccount would not always have been set, and any existing value could be left over from a previous payment.

This removes paymentGatewayAccount as a passed in property, and relies on it being set in onboarding, before being stored as state (in the use case) much like the connected reader is.

A follow up to this change would be to remove this stored state and pass the values to the various tracking functions, however doing that in this branch made the change too large.