8773 Reveal TTP feature toggle in release builds
This commit makes Tap to Pay on iPhone available for merchants to opt-in, using the experimental feature toggle, on an App Store build.

The toggle is only available on phones, as TTP is not supported on iPad.

The Tap to Pay on iPhone entitlement is not available for Enterprise certificates, so we can’t enable the feature toggle in our Alpha build, which is used by CI to build the enterprise distribution of the app. Since this is an uncommon configuration for a feature flag, I’ve called that out in the comments.