Rewrite some tests with neater `Data("*".utf8)`
That is instead of `"*".data(using: .utf8)!`.
This approach is better because it avoids the force unwrapping.
Even though it was safe to do, having to use it it's still a bit of a
smell, or at least something that would give pause to a reader and made
them as "why?".

See
https://github.com/wordpress-mobile/WordPressAuthenticator-iOS/pull/740#discussion_r1107526775