8992 Arrange tests of firstTransaction/powerUser surveys
Previously we did not inject the CardPresentPaymentsConfiguration, so it defaulted to an unsupported country and these tests recieved no banner in the response.

Also, we did not pass in the storageManager, so none of the orders we created were available to the code under test.

These tests still fail at this commit, because the logic under test (incorrectly) requires them to have COD enabled.