8992 Only require COD for the first survey
The inPersonPaymentsCashOnDelivery survey is intended to survey merchants who sell with COD (i.e. may be interested in IPP) but with a non-WCPay solution. This is supposed to target people who may switch to IPP with WCPay.

The other two surveys should not be limited to COD merchants, because not all WCPay merchants will want to offer that as a payment option on their website, even if they do collect IPP in some circumstances.

This also adds tests for the negative of this scenario, and when the store is in an unsupported country.