8992 Require receipt_url for all orders wrt IPP surveys
All the IPP feedback survey logic should be based on WCPay IPP transactions. Previously, the results controllers did not include the required filter for the `receipt_url` metadata, so we (mostly) considered web orders as if they were IPP orders.

The exception to that was a post-fetch filter on the recentTransactions results.

This commit moves the filter to the predicate, and uses the predicate for both the results controllers (all WCPay IPP transactions, and recent WCPay IPP transactions.)

We never consider web orders, and now we can’t miss the filter out of our logic by mistake.