8992 Push the feedback banner to the viewcontroller
Previously we pushed the instruction to show a feedback banner to the view, and then the view _pulled_ the banner type it should show. This was a little confusing and while it worked, it’s error prone.

This change pushes the banner tyep along with the instruction to show the banner.