8992 Improve display of feedback banners
Feedback banners were not reliably shown. This… isn’t the prettiest, because we don’t have a WCPayIPP related action to use as the prompt to update the banner type, so it uses the order action’s completion handler as a bit of a side-effect.

It’s not too bad, because the order action is the one that fetches the relevant data, and it’s only temporary anyway.