8992 Show OrderCreation banner when no IPP survey
By allowing `none` for the IPP banner source, we accidentally prevented the OrderCreation banner from showing, as we could “show” an empty IPP banner. This didn’t appear on screen, but dropped out of the top banner presentation code before the order creation banner got a chance to show.

This change removes the optionality from the TopBanner enum, so we either have an IPP banner to show, or we don’t (and we move to Order Creation), no half-way house.

This also fixes an issue where an (incorrectly shown) COD banner was not always cleared when all the data was fetched.