8774 Remove Tap to Pay toggle
This commit turns Tap to Pay on iPhone on for all merchants.

We still check for device support using the method which Stripe expose, called from `CardPresentPaymentPreflightController` using `CardPresentPaymentAction.checkDeviceSupport`. This checks for iOS 16+ and the required hardware, and we only show the selection modal if it returns `true`.

This change deliberately leaves the legacy code in place, because of other ongoing work and the chance that we may revert this in the 12.6 beta period. Leaving the legacy code in makes that potential reversion more straightforward. We will remove it once 12.6 is released.