9010 Replace use of ZendeskUI func for constraints
I used `constrainToSuperViewEdges` for the table view when adding it to the CardReaderSettingsConnectedViewController’s view. This was a new requirement after removing the storyboard for this view, which was previously responsible for instantiating the table view.

Unfortunately, I didn’t notice that this convenient method was actually from Zendesk’s `CommonUI` framework, which was not appropriate to use in this context.

This commit moves to more traditional layout constraint use.