Pass salectedItemIDs to ProductSelectorViewModel
In order to refresh the list of selectedItemIDs each time we open the product selector, we need to re-initialize the ProductSelectorViewModel, so it can’t be lazy. This is necessary to keep the state of selected products if these are removed directly from the Order, rather than from the product selector.