Point WordPressAuthenticator to version `~> 5.5` in the `Podfile`
Worth noting that 82ee192a42e0c48a7928a9c58bfeaeaa5d03be41 reported to
do that already, but it was an update only at the `Podfile.lock` level.

You might also notice that commit pointed the pod to
d7445ad42b6a5774eed45b394810881d5c01a7c6 which is not the 5.5.0 commit.
However, that's the commit of the `5.5.0` merge into the library's `trunk`,
so it's effectively the same.