9010 Show spinner on Set Up button when started
The PaymentGatewayAccounts are fetched before any alert is shown by the ConnectionController. This means there can be a small gap between tapping the `Set up Tap to Pay on iPhone` button and showing any change in the UI.

Here, we add a spinner to the button, so when it’s pressed the user sees that the app is loading from the network. The button is returned to its original state when the ConnectionController completes, success or failure.