9010 Fix auto-disconnect of bluetooth readers
If you went in to the Set up Tap to Pay on iPhone screen, then dismissed it and later connected to an external card reader, the reader you connected to would automatically disconnect as soon as it had completed connecting.

This was due to the SetUpTapToPayInformationViewModel being held in a retain cycle. That VM observes card reader connections, and automatically disconnects from external readers when found, so the merchant can get on with setting up tap to pay.

Unfortunately, the alerts presenter held its root view, which held the view model, which held the alert presenter – and that’s a retain cycle, so the observation of connected readers never stopped.

This commit breaks the retain cycle at the alert presenter.

Additionally, there were closure-capture retain cycles in both the Manage Card Readers screen, and the Set Up screen, due to not using [weak self] in certain closures which were passed from the view controller to the view. Since the view controller has a reference to the view, this also created retain cycles.