9278 Hide/disable Set up Tap to Pay row as needed
Currently (temporarily) the TTP row will be hidden if the device does not support TTP, or the store does not support TTP.

If the row is shown, it will be disabled until onboarding reaches `.complete`.

This behaviour will change when we add support for universal links to the set up flow.