9370 Show onboarding before TTP set up flow
When onboarding is incomplete, a merchant will not be able to complete the set up flow. Up to now, we’ve prevented the flow from being opened until onboarding is complete, relying on the in-line onboarding offered on the Payments menu.

With the addition of the Universal Link to the Set up Tap to Pay on iPhone flow, that’s no longer sufficient, as we don’t want to make a merchant wait on the menu screen before completing the navigation from their tap on the link.

This change removes the block on opening the set up flow, and shows the onboarding flow when required before set up begins.