Use passed currency on receipts
UK In-Person Payments receipts were showing with $ signs.

This was due to the currency code used from the API being `gbp` instead of `GBP`, which meant it didn’t match the `CurrencyCode.rawValue` when checked.

All our `CurrencyCode` values are upper case, so it seems reasonable to convert anything that’s passed in to upper case as well, to avoid this issue in future.