9370 Add tests for IPP Onboarding shouldShow logic
When the In Person Payments Onboarding screens are presented within a `PaymentSettingsFlow`, e.g. before Set up Tap to Pay on iPhone, the logic for setting `shouldShow` and calling `didChangeShouldShow` determines whether the overall onboarding screen is displayed or not.

This adds tests for this logic. When the `shouldShow` tri-state `isTrue`, and it’s the first screen in the flow saying it `shouldShow`, the `InPersonPaymentsViewController` will be displayed. We want to return `isTrue` when there’s a loading or error status in Onboarding, and `isFalse` when it’s completed.