Replace “Built in reader” with “Tap to Pay on iPhone”
In user-facing text, Apple and Stripe have asked that we refrain from using the term “reader” or “built-in reader”, instead referring to Tap to Pay on iPhone.

This language was used in some of the connecting modals and error messages.

This change replaces these names as required, and reduces the use of `builtInReader` in the code as well (though, it’s still part of the Stripe SDK terminology so I’ve not removed it from all parts of the code, mostly comments.)