Use SPM cache in all Buildkite steps
I thought we could get away with it in the "test without building"
steps, but evidently that's not the case because CI fails.

See
https://github.com/woocommerce/woocommerce-ios/pull/9623#issuecomment-1535805390

At this point, I prefer the overhead of downloading the cache when it
might be skipped to the additional work of looking into the guts of
Fastlane to see if there's a way to work around the error seen in CI.
Downloading the SPM cache is anyways better that downloading all the
packages during resolution.