9735 Add basic Tap to Pay reconnection controller
When we launch or foreground the app, for merchants who have used Tap to Pay in the past, we should automatically reconnect the built in reader in the background.

This is to save time when taking a payment, and is recommended by Stripe and Apple. The connection is usually reasonably fast (7s in my experience) but that can feel like a long time when you’re taking a payment.