9735 Adopt in-progress Tap to Pay reconnections
When we launch or foreground the app, for merchants who have used Tap to Pay in the past, we automatically reconnect the built in reader in the background.

If the merchant starts a card payment during that time, we need to start showing the alerts from the connection process. It’s not possible to cancel a connection, so if they choose a hardware reader, we wait for the reconnection to complete, then immediately disconnect from the tap to pay reader, and connect to the hardware reader instead.