9735 Check onboarding is complete before reconnect
A connection attempt when onboarding is not `complete` will fail. By checking the onboarding cache for `complete` status, we reduce the network traffic and number of errors shown for onboarding failures which the user won’t ever see.

Since we can be fairly confident the reconnection will fail when we don’t have `complete` as the cached status, we don’t need to try to reconnect.