9735 Prevent stale reader use after store change
With the background automatic reconnection to Tap to Pay, it became possible to switch stores while a card reader connection was in progress.

The `reset` action is called when we switch stores – it  disconnects from connected readers, then cleans up the config.

However, with Stripe’s SDK it’s not possible to cancel an _in-progress_ connection. With this change, we invalidate any in-progress connection, and immediately disconnect the reader if it succeeds.