9773 Prevent TTP reconnection if no location authn
We reconnect to the reader automatically on foregrounding the app, if the device and store support it and TTP has been used before.

For users which grant `allow once` permissions, this means that in the next app session, they’ll be asked to provide location permission without the context of a payment attempt, which is strange – the system permission alert pops up over whatever screen they happen to be on, interrupting their flow.

With this change, we only attempt to reconnect if we already have location permission – this means we no longer show a location permission request out of context.