9884 Fix UpgradesViewModel tests to use mock
The mock IAPManager wasn’t used by all the tests, and when we call `sut.fetchPlans` without injecting the mock, `inAppPurchasesAreSupported` gets called and leaks its continuation. This may be related to using a testing Stores manager and dispatching actions to it, but I didn’t dig in to that.

This change avoids us testing any of these methods on the real IAP manager.

Additionally, we’ve made the injection a bit clearer by using obvious test data.