9909 Add feature flag for hardcoded upgrade details
We’ll provide hardcoded upgrade details for M1 of in-app purchases for Woo Express plans.

If these details change in future, we’ll want to be able to invalidate the file where they’re held and prompt the user to check the web and/or update their copy of the app, instead of continuing to show out of date info.

This adds a remote feature flag to the app which we can use to check whether the details have been invalidated.