9909 Hide hardcoded features when remote flag off
We added a remote flag to allow us to invalidate the hardcoded data regarding Woo Express plans.

We assume it’s `true` until we fetch `false`, at which point we hide the feature details, and replace them with a button to open the express page in a web view. This is a fallback only, and we’ll remove it once we can fetch the feature details from an API