9878 Move to the `waiting` state when IAP closes
There’s no particularly good signal from Apple for when the In App Purchase drawer closes – the best we have is a `didBecomeActiveNotification` with no distinguishing features in its user info.

This commit adds a closely scoped observation of that notification, and advances the view to `.waiting` when it’s received, as long as the view is still in `.purchasing(_)`, i.e. we’ve not had a cancellation or error or other result from the purchase since we started.

We aggressively stop this observation, because other more concrete results are better indicators, and we don’t need to speculatively advance to the waiting state if we get those. Unfortunately (but understandably, given the lifecycle), the notification usually arrives before the other results, so we delay our handling of the notification by half a second to let everything else catch up.

None of this is ideal, but it works, and it’s what we have.