10075 Show error when IAP transaction already used
Initially, our IAP plan purchases only support upgrading a single site. Upgrading multiple sites is prevented by checking whether the merchant has already got an active entitlement to an In-App Purchase, and showing the Maximum sites upgraded error at the start of the Upgrade flow if they have.

Unfortunately, if the MobilePay server and Apple are out of sync, it's possible to make an In-App Purchase to upgrade a second site which will succeed with Apple (i.e. taking the merchants money), fail with MobilePay, but be announced as success in the app. The IAP Transaction will be marked as finished, (i.e. we're telling Apple that we've furnished the merchant with the upgrade they paid for) and the site will remain on the trial. We won't ever show the merchant an error.

We do get a 422 error in this situation, but previously, we ignored it.

422 errors are common in our implementation, as we get lots of renewal transactions which we have to handle on app start up – when we send these to the API, MobilePay already knows about them because of the server to server notifications from the App Store Connect API, so sends us a 422. We mark the transaction as finished and ignore the error, which is correct for renewals.

Unfortunately, we can't distinguish (using the transaction) between these renewals, and new purchases, because the `reasonStringRepresentation` property which Apple documents... doesn't seem to exist. Perhaps we can use the new `reason` property once iOS 17 reaches wider adoption.

## Mitigation
In this commit, we reduce the impact of this issue by surfacing any 422 errors that happen during the `purchaseProduct` flow, as opposed to being passed to the app on the transactions queue, at start up etc. We’ve still taken the merchants money without upgrading their site, but at least we are telling them that there’s an issue and that they need to contact support.