Handle pending transactions
Strong Customer Auth or Parental controls can result in an In App Purchase returning `.pending` as the transaction state, then later returning the successful transaction on `Transaction.updates`.

While the app correctly handles these results when they’re recieved on the updates queue, previously, the app would have sat on the `Waiting` screen forever in this case.

This PR updates our InAppPurchaseStore so that it
1. keeps the completion handler around until the pending result is confirmed, and
2. does not call the completion handler with a pending result (because we can only call it once.)

Unfortunately, Apple does not update us to tell us about a `.declined` transaction, so the app will still be in the waiting state forever if the parental consent is declined or SCA is failed. For that reason, the screen can be dismissed and the upgrade tried again, as recommended by Apple.