10023 Track upgrade dismissal centrally
This isn’t the ideal way to track dismissal of these screens, but I can’t find a better one which takes account of swipe to dismiss, as long as we’re using a HostingController.

The ViewModel probably needs to provide a function for `didInteractivelyDismiss`, which is then made available by the HostingController, and called from UIKit when the ViewController/TransitionDelegate handles a swipe. But this does the trick, for the most part, and is a lot quicker.