10156 Fix print shipping labels from order search
Previously, if you opened an order from the search results, it was not possible to print any shipping labels for that order.

This was due to the behaviour of `UIPrintInteractionController.present` when called before `dismissPrintInProgress` is finished. When called above a SearchUIController, the print dialog was being presented from the in-progress view, but not until _after_ it had been removed from the view heirarchy.

This issue did not affect the printing process when the Order Details were opened directly from OrderListViewController, without using search.

This commit waits for the dismissal to complete before calling `present`, which resolves the timing issue.