10286 Disable plan selection while purchasing
It takes a little while after tapping `Purchase` for Apple’s In-App Purchase drawer to show up.

Previously it was possible to change your selected plan after tapping Purchase, while waiting for the IAP drawer to show, but any changes would not be reflected when the draw did appear. This was potentially confusing.

This disables plan selection and opening the all features view while waiting for Apple.