Prevent reset of the  UpgradesView on Purchase tap
Previously, when the merchant tapped `Purchase Essential Monthly` or similar, we sent a new view state, `purchasing(plan, [plans])`, which set the purchasing state (i.e. spinner on the Purchase button, disabling the plan selection).

Unfortunately, this also resulted in a full view redraw, which means we lost the user’s place. If they’d selected a monthly plan, it would be hidden as the view moved back to the default yearly plans section, and any scroll position would be lost, which looked a little strange.

Here, we replace the `purchasing` state with a published property for `isPurchasing` on the view model, which is bound by the `OwnerUpgradesView` where needed. This value also lets us check whether we should advance to the waiting state after the IAP drawer is dismissed – if we’re no longer purchasing because of an error or cancellation, we shouldn’t show the waiting state.