Enable Google SDK-less login in production
We have a system for remote feature flags, which would have been my
preferred option, but we cannot use it here because the system to get
them is asynchronous while our value needs to be available shortly after
the app launch.

Things I tried before this:

- Update the callers to be `async`. This resulted in an exploding number
  of required changes up the call-chain.
- Wrap the remote flag `async` in `Task { @MainActor }`. This failed
  because code depending on `WordPressAuthenticator` being initialized
  run together with the async call and crashed.

The SDK-less sign in has not given us trouble in Jetpack and WordPress.
It should be safe enough to hardcode the value here.