Show monthly plans before yearly plans
After releasing M2 of IAP, which added the yearly plans, our sales dropped to zero for ten days.

This may be due to the sticker shock of a $300-539 In-App Purchase, compared to a $39/month monthly plan.

We’ve swapped the order and start page to show the monthly prices first to see what impact that has on sales.