Use ServiceLocator for stores because of lifecycle
Since WooAnalytics is also fetched from the ServiceLocator, it can’t use a default value for `stores` when init is called that also uses the ServiceLocator, as this is not intialised yet when called on app start.

This commit resolves that issue by going back to using ServiceLocator for the dependency at each call site, and setting the mock in tests as required.