Show announcements when current version in range
Announcements are for a particular version, but also get a max/minimum version range where they should show.

If the current version is in the range, the announcement will have shown on launch. These users may also want to look at the announcement again, after dismissing it.

We should show the What’s New row if there’s an announcement for which the current version falls within the range.