Reuse payment intent for appropriate retries
To avoid issues with duplicate payments, retries of payments after the payment intent has been created, should be done using the same payment intent.

Here we add a new `presentRetryWithoutRestarting` method which can be used in these circumstances to reuse the active payment intent. In order to make this work without being repetitive with the calls to `attemptPayment` and `retryPayment`, we now hold the handlers in the PaymentOrchestrator for the duration of the payment.