Retry requiresCapture PIs when there’s an error
These Payment Intents are good to use. We may have lost some context, e.g. by the `confirm` request that the SDK makes from `processPayment` being lost due to bad network conditions.

When that happens, we should continue with capture, and let WooPayments return an error if the payment fails when it attempts capture.